Add Certification Station and book, graphics controller, abstract station.
