+++ Assets/Prefabs/Station.prefab
@@ -9,7 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 4364338940078858265}
-  - component: {fileID: 4364338940078858267}
+  - component: {fileID: 4524239868664230309}
   m_Layer: 0
   m_Name: Station
   m_TagString: Untagged
@@ -38,7 +38,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &4364338940078858267
+--- !u!114 &4524239868664230309
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -47,7 +47,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 4364338940078858264}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 973b6b70efcaaf94aac4bc60458b0f09, type: 3}
+  m_Script: {fileID: 11500000, guid: 0e7899e951f09474cb5e07ee17d35b9e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   chairs:
