+++ Assets/Scripts/Components.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 073f3d9ba702a6342b8ead310e72c2ad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
