+++ Assets/Scripts/Components/CertificationStation.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+public class CertificationStation : Station
+{
+    [SerializeField]
+    GraphicsController bookGraphics;
+
+
+}
