+++ Assets/Scripts/Components/CertificationStation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b27d690ee28cc5140ae62878405d17b6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
