+++ Assets/Scripts/Components/GraphicsController.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+public class GraphicsController : MonoBehaviour
+{
+    List<GameObject> graphics;
+
+
+}
