+++ Assets/Scripts/Components/WorkStation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e7899e951f09474cb5e07ee17d35b9e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
